requirements.txt
ftfy==6.3.1
numpy==1.24.3
opencv_python==4.10.0.84
pandas==3.0.2
Pillow==12.2.0
regex==2023.8.8
scipy==1.17.1
timm==1.0.26
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
yacs==0.1.8
